pyproject.toml
[tool.poetry]
name = "forecast-2-0-pipeline"
version = "1.0.0"
description = "Pipeline ETL pour l'intégration de données météorologiques - GreenAndCoop"
authors = ["Alhassane AHMED <ali75009@gmail.com>"]
readme = "README.md"
license = "Proprietary"

# Indiquer les packages Python
packages = [
    { include = "pipeline", from = "src" },
    { include = "loaders", from = "src" },
    { include = "utils", from = "src" },
    { include = "scripts", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.4"
numpy = "^1.26.2"
boto3 = "^1.34.10"
botocore = "^1.34.10"
pymongo = "^4.6.1"
dnspython = "^2.4.2"
openpyxl = "^3.1.2"
xlrd = "^2.0.1"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
requests = "^2.31.0"
urllib3 = "^2.1.0"
python-dateutil = "^2.8.2"
pytz = "^2023.3"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
tqdm = "^4.66.1"
jsonschema = "^4.20.0"
python-json-logger = "^2.0.7"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
faker = "^21.0.0"
pytest-asyncio = "^0.21.1"
black = "^23.12.1"
flake8 = "^7.0.0"
pylint = "^3.0.3"
mypy = "^1.8.0"
isort = "^5.13.2"
types-requests = "^2.31.0"
types-python-dateutil = "^2.8.19"
pandas-stubs = "^2.1.1"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"

[tool.poetry.scripts]
forecast-pipeline = "main:main"
transform-mongodb = "scripts.transform_to_mongodb:main"
migrate-mongodb = "scripts.migrate_to_mongodb:main"
mongodb-crud = "scripts.mongodb_crud:main"
latency-report = "scripts.query_latency_report:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-vv -ra"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
